+++ src/views/Admin/Settings_Pages/SettingDashboard.vue
@@ -21,11 +21,11 @@
             MODIFY USER ROLES
           </button>
         </router-link>
-        <router-link to="/adminSettings">
+        <router-link to="/adminSettings/ChangeBackground">
           <button
-            class="btn btn-primary ms-3"
+            class="btn btn-outline-success ms-3"
             style="
-              background-color: #135c1f;
+              /* background-color: #135c1f; */
               font-size: 12px; /* Adjust the font size as needed */
               padding: 10px 30px 10px 30px;
             "